feat: studyList 조회 API
