+++ src/lib/constants/Url.ts
@@ -5,6 +5,7 @@ export const EMAILAUTHCODE_URL = '/api/auth/email/confirm'
 export const REFRESHTOKEN_URL = '/api/auth/refresh'
 export const EDIT_MEMBER_URL = '/api/member/edit'
 export const GET_REFRESH_URL = '/api/auth/refresh'
+export const STUDY_URL = '/api/study'
 
 export const GITHUB_LOGIN_URL =
   'https://github.com/login/oauth/authorize?client_id=75cd6db13c045cad11e6&redirect_uri=http://localhost:3000/callback'
