+++ src/lib/constants/constants.ts
@@ -2,3 +2,5 @@ export const USER_KEY = 'user'
 
 export const SET_HEADER_VISIBILITY = 'SET_HEADER_VISIBILITY'
 export const SET_USER_MENU_VISIBILITY = 'base/SET_USER_MENU_VISIBILITY'
+
+export const STUDIES_PER_PAGE = '8'
